[Refactor] 코드 정리 및 파일들 정리
