+++ public/data/common/bannerInfo.json
