+++ public/data/common/promotionInfo.json
