+++ src/Router.js
@@ -6,7 +6,6 @@ import Main from "./pages/Main/Main";
 import Esg from "./pages/Menu/Esg";
 import RenterCar from "./pages/Menu/RenterCar";
 import Accomodation from "./pages/Menu/Accomodation";
-import Food from "./pages/Menu/Food";
 import Mypage from "./pages/Mypage/Mypage";
 import RentCar from "./pages/RentCar/RentCar";
 import Preparing from "./pages/Menu/Preparing";
@@ -17,7 +16,7 @@ import SignupFin from "./components/signup/signupFin";
 import FooterTop from "./components/Footer/FooterTop";
 import FooterBottom from "./components/Footer/FooterBottom";
 import Restaurant from "./pages/Restaurant/Restaurant";
-import RestaurantDetail from "./pages/Restaurant/RestaurantDetail";
+import RestaurantDetail from "./pages/Restaurant/RestaurantDetail/RestaurantDetail";
 import RestaurantMain from "./pages/Restaurant/RestaurantMain";
 
 function Router() {
@@ -29,7 +28,9 @@ function Router() {
         <Route path="/esg" element={<Esg />} />
         <Route path="/rentercar" element={<RenterCar />} />
         <Route path="/accomodation" element={<Accomodation />} />
-        <Route path="/food/:id" element={<Food />} />
+        <Route path="/restaurant" element={<RestaurantMain />} />
+        <Route path="/restaurant/:id" element={<Restaurant />} />
+        <Route path="/restaurant/detail/:id" element={<RestaurantDetail />} />
         <Route path="/preparing" element={<Preparing />} />
         <Route path="/mypage" element={<Mypage />} />
         <Route path="/rentercar/rentcar/:id" element={<RentCar />} />
@@ -38,9 +39,6 @@ function Router() {
         <Route path="/signup" element={<Signup />} />
         <Route path="/signupFin" element={<SignupFin />} />
         <Route path="/preparing" element={<Preparing />} />
-        <Route path="/restaurant" element={<RestaurantMain />} />
-        <Route path="/restaurant/:id" element={<Restaurant />} />
-        <Route path="/restaurant/detail/:id" element={<RestaurantDetail />} />
       </Routes>
       <FooterTop />
       <FooterBottom />
