+++ src/components/Car/Car.js
@@ -1,7 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { Swiper, SwiperSlide } from "swiper/react";
 import SwiperCore, { Pagination } from "swiper";
-import { FaCar } from "react-icons/fa";
 
 import "./Car.scss";
 import "swiper/css";
@@ -13,7 +12,7 @@ function Car() {
   const [data, setData] = useState([]);
 
   useEffect(() => {
-    fetch("/data/carInfo.json")
+    fetch("/data/common/carInfo.json")
       .then((res) => res.json())
       .then((data) => {
         setData(data.carInfo);
@@ -22,10 +21,7 @@ function Car() {
 
   return (
     <div className="content-wrapper">
-      <h1 className="title">
-        최저가 렌터카 추천
-        {/* <FaCar /> */}
-      </h1>
+      <h1 className="title">최저가 렌터카 추천</h1>
       <Swiper
         spaceBetween={20}
         slidesPerView={3}
